README.md
This is my very first project. It's from The Odin Project. I'm trying to build a page that will look similarly like Google home page, it is supposed to have all the elements that Google.com has. I should demonstrate my modest begginer's knowledge of the basics of using html and css.